// Source/Microsoft.Teams.Apps.EmployeeTraining/ClientApp/src/models/language-direction.ts
// <copyright file="language-direction.ts" company="Microsoft">
// Copyright (c) Microsoft. All rights reserved.
// </copyright>

export enum LanguageDirection {
    /** Indicates that the language direction is right-to-left*/
    Rtl = "rtl",

    /** Indicates that the language direction is left-to-right*/
    Ltr = "ltr",

    /** Indicates that the language direction is auto*/
    Auto = "auto",
}